pdf-backend/api/requirements.txt
fastapi
uvicorn
pytesseract
pillow
playwright
opencv-python
python-multipart